fix(query,cli,mcp): one implementation of task file selection
`drift prepare` and `get_task_preflight` each had their own copy of "which files
matter for this task". The copies had diverged three ways, all measured:

1. THE IN-SCOPE PREDICATE. The CLI tested `path_globs` directly while the enforcement
   path uses `conventionScopeFiles`, so `prepare` told an agent a file was "in scope
   for convention_X" using a rule `check` does not use to enforce convention_X. BB-11
   already adjudicated glob-only as a bug rather than a policy - and fixed only the MCP
   copy. On dub's real file list the two answer differently by 623 vs 171 files for a
   contract with exclusions, and 0 vs 4,091 for one with empty globs.

2. GITIGNORE. The CLI honoured it. The MCP copy never did.

3. DECLARATION FILES. When the scan learned to record `.prisma` (#110), only the CLI
   walker learned it. Measured on a fixture: for "add a field to the prisma schema
   model Thing", `prepare` returned prisma/schema.prisma and `get_task_preflight` did
   not. That one was introduced by #110 and missed because only the CLI path was
   exercised.

The MCP server cannot import from the CLI - the boundary check enforces it - which is
why a second copy existed at all. So the implementation moves to `@drift/query`, the
home the boundary comment already designates for shared logic: "where T51 moved
preflight ranking after the two copies silently diverged." T51 fixed ranking and left
selection duplicated. This is the other half.

The canonical predicate wins on scope and the gitignore-aware walk wins on discovery,
because each was the side already adjudicated correct. The walk moves too - leaving it
in the CLI is what forced the copy that caused divergences 2 and 3 - so there is now
one `walkIndexableFiles`, re-exported to the CLI's five call sites.

Verified: CLI and MCP now return byte-identical results for the same repo and task.

Caught while doing this: my first pass rewrote `tokenizeTask` from assumption rather
than from either implementation, silently changing the split to drop `_`, `/` and `-`
and raising the length floor from 3 to 4. The two real implementations were
byte-identical, and that would have been a behaviour change disguised as a refactor -
it decides which files earn a `task token:` reason, which is what puts a file in front
of an agent at all. The shared copy is now verbatim, with the tokenizer pinned by a
test that asserts `app/api` stays ONE token.

pnpm verify:ci exits 0 reaching beta:proof. Determinism unchanged (f208ce7ce3bef7a0).
